+++ originacion/src/main/java/com/banquito/sistema/originacion/repository/VehiculoRepository.java
@@ -4,5 +4,5 @@ import com.banquito.sistema.originacion.model.Vehiculo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface VehiculoRepository extends JpaRepository<Vehiculo, Integer> {
-    boolean existsByIdentificadorVehiculoId(Integer idIdentificadorVehiculo);
+    boolean existsByIdentificadorVehiculoId(Long idIdentificadorVehiculo);
 }
